enterprise-view finalizada a falta de pulir detalles

+++ src/main/java/com/cesur/aplicaciondual/domain/entities/empresa/EmpresaDAO.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 
 public interface EmpresaDAO {
     public ArrayList<Empresa> getAll();
-    public Empresa get(Long id);
+    public Empresa get(Integer id);
     public Empresa save(Empresa data);
     public void update(Empresa data);
     public void delete(Empresa data);
